requirements.txt
# Core
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pyyaml>=6.0.1
lxml>=5.0.0

# Playwright for JS-rendered sites
playwright>=1.40.0

# LLM providers — install only what you use
anthropic>=0.40.0
openai>=1.50.0
google-generativeai>=0.8.0

# Vector store (for future FAISS RAG layer)
faiss-cpu>=1.8.0
numpy>=1.26.0
